requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-crispy-forms==2.1
django-widget-tweaks==1.5.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
whitenoise==6.6.0
